feat: use codeset resources
