+++ src/api/endpoints/index.ts
@@ -27,3 +27,7 @@ export const TESTBED_API_BASE_URL = process.env.REACT_APP_TESTBED_API_BASE_URL
 
 export const EXT_REGISTRATION_SERVICE_URL =
   process.env.REACT_APP_EXT_REGISTRATION_SERVICE_URL || 'http://localhost:3001';
+
+export const CODESETS_BASE_URL = process.env.REACT_APP_CODESETS_BASE_URL
+  ? removeTrailingSlash(process.env.REACT_APP_CODESETS_BASE_URL)
+  : 'http://localhost:3166';
