+++ src/api/index.ts
@@ -1,4 +1,5 @@
 import * as auth from './services/auth';
+import * as codesets from './services/codesets';
 import * as consent from './services/consent';
 import * as data from './services/data';
 import * as user from './services/user';
@@ -8,6 +9,7 @@ const api = {
   user,
   data,
   consent,
+  codesets,
 };
 
 export default api;
