+++ src/api/services/user.ts
@@ -2,10 +2,7 @@ import axiosInstance from '../axiosInstance';
 
 // types
 import {
-  UserProfile,
-  CountryOption,
-  OccupationOption,
-  LanguageOption,
+  UserProfile
 } from '../../@types';
 
 // endpoints
@@ -21,37 +18,4 @@ export async function get() {
 
 export async function patch(payload: Partial<UserProfile>) {
   return axiosInstance.patch(`${USERS_API_BASE_URL}/user`, payload);
-}
-
-// profile related codesets
-export async function getCountries(): Promise<CountryOption[]> {
-  const { data } = await axiosInstance.get(
-    `${USERS_API_BASE_URL}/code-sets/countries`
-  );
-  return data;
-}
-
-export async function getOccupations(): Promise<OccupationOption[]> {
-  const { data } = await axiosInstance.get(
-    `${USERS_API_BASE_URL}/code-sets/occupations`
-  );
-  return data;
-}
-
-export async function getOccupationsFlat(): Promise<OccupationOption[]> {
-  const { data } = await axiosInstance.get(
-    `${USERS_API_BASE_URL}/code-sets/occupations-flat`
-  );
-  return data;
-}
-
-export async function getLanguages(): Promise<LanguageOption[]> {
-  const { data } = await axiosInstance.get(
-    `${USERS_API_BASE_URL}/code-sets/languages`
-  );
-  return data;
-}
-
-export function getGenders() {
-  return axiosInstance.get(`${USERS_API_BASE_URL}/code-sets/genders`);
 }
